[feat] add user rating info & merge Frontend/KYR/piece_of_cake into Frontend/POC

+++ BackEnd/common_pjt/src/main/java/com/E203/pjt/model/dto/req/UserReqVO.java
@@ -14,7 +14,8 @@ public class UserReqVO {
     private String userNickname;
     private String userPassword;
     private String userImage;
-    private Integer userRating;
+    private Double userRating;
+//    private Integer userRating;
     private String userLat;
     private String userLng;
     private String userAccount;
